cleaned some conversion warnings

+++ src/bmdbg.h
@@ -145,7 +145,7 @@ int read_dump_file(const std::string& fname, VT& data)
     fin.seekg(0, std::ios::end);
     fsize = fin.tellg();
     
-    data.resize(fsize/sizeof(value_type));
+    data.resize(unsigned(fsize)/sizeof(value_type));
 
     if (!fsize)
     {
@@ -767,7 +767,7 @@ int file_save_svector(const SV& sv, const std::string& fname, size_t* sv_blob_si
         return -1;
     }
     const char* buf = (char*)sv_lay.buf();
-    fout.write(buf, sv_lay.size());
+    fout.write(buf, unsigned(sv_lay.size()));
     if (!fout.good())
     {
         return -1;
